create endpoint for si-business and show all Mesin

+++ SI-FACTORY/src/main/java/apap/group/assignment/SIFACTORY/model/PegawaiModel.java
@@ -28,7 +28,7 @@ import java.util.List;
 public class PegawaiModel implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idPegawai;
+    private Long idPegawai;
 
     @NotNull
     @Size(max=50)
